refactored the image reading and profile pic uploading parts to be separate
