+++ client/src/components/profile-pic-uploader/profile-pic-uploader.component.jsx
@@ -1,34 +1,20 @@
 import React from 'react';
 import Client from '../../utils/HTTPClient';
 import { Button } from '@material-ui/core';
+import ImageReader from '../image-reader/image-reader.component';
 
 const ProfilePicUploader = ({ socket, onCompleted }) => {
-  const handleChange = e => {
-    const input = e.target;
-
-    const reader = new FileReader();
-    reader.onload = () => {
-      const dataUrl = reader.result;
-      const pic = { name: input.files[0].name, data: dataUrl };
-      // TODO: send picture
-      Client.request('/user/avatar', 'POST', pic).then(res => {
-        const { pic, userId } = res;
-        socket.emit('updateProfilePic', { userId, profilePic: pic });
-        onCompleted();
-      });
-    };
-    reader.readAsDataURL(input.files[0]);
+  const onRead = pic => {
+    Client.request('/user/avatar', 'POST', pic).then(res => {
+      const { pic, userId } = res;
+      socket.emit('updateProfilePic', { userId, profilePic: pic });
+      onCompleted();
+    });
   };
-
   return (
     <Button variant="contained" component="label">
       Upload Profile Picture
-      <input
-        type="file"
-        accept="image/x-png,"
-        onChange={handleChange}
-        style={{ display: 'none' }}
-      />
+      <ImageReader style={{ display: 'none' }} onRead={onRead} />
     </Button>
   );
 };
